% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rev_reaction.R
\name{add_rev_reaction}
\alias{add_rev_reaction}
\title{completes all reversible reactions in lipid species biosynthetic network}
\usage{
add_rev_reaction(network_edge, species_net)
}
\arguments{
\item{network_edge}{A data.frame recording edge information of the reference
lipid biosynthetic network in iLipdiome.}

\item{species_net}{Lipid species biosynthetic network. Output of
"build_species_net".}
}
\value{
Lipid species biosynthetic network with complete reversible reactions
}
\description{
completes all reversible reactions in lipid species biosynthetic network
}
