% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_char_table.R
\name{build_char_table}
\alias{build_char_table}
\title{Transforms lipid expression table}
\usage{
build_char_table(raw_data, network_node)
}
\arguments{
\item{raw_data}{A data.frame where rows are lipid species and columns are
samples. Lipid names should be provided in the first column. Lipid values
should be numeric.}

\item{network_node}{A data.frame recording iLipidome-supported lipid classes
and related information.}
}
\value{
list containing selected lipid expression table and selected lipid
characteristics table
}
\description{
Transforms lipid expression table (“exp”) into two iLipidome inputs:
selected lipid expression table (“exp_sel”) and
selected lipid characteristics table (“char_sel”).
}
