% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_FA_net.R
\name{build_FA_net}
\alias{build_FA_net}
\title{Fatty acid biosynthetic network transformation}
\usage{
build_FA_net(FA_network, unprocessed_data_result)
}
\arguments{
\item{FA_network}{A data.frame describing the reference FA biosynthetic
network in iLipdiome.}

\item{unprocessed_data_result}{FA_network: A data.frame describing the
reference FA biosynthetic network in iLipdiome.}
}
\value{
transformed fatty acid network
}
\description{
Fatty acid biosynthetic network transformation
}
