% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_species_net.R
\name{build_species_net}
\alias{build_species_net}
\title{Construct lipid biosynthetic network}
\usage{
build_species_net(species_substructure)
}
\arguments{
\item{species_substructure}{Output of "species_sub_transform"}
}
\value{
lipid species biosynthetic network
}
\description{
Uses species substructures to construct lipid biosynthetic network
}
