% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FA_network}
\alias{FA_network}
\title{Dataset FA network}
\format{
dataframe with x rows and y vars
\describe{
\item{row1}{description 1}
\item{row2}{description 2}
}
}
\source{
{URL or created in house}
}
\usage{
FA_network
}
\description{
Contains FA network data
}
\examples{
data(FA_network)
}
\keyword{datasets}
