% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_sub_extract.R
\name{FA_sub_extract}
\alias{FA_sub_extract}
\title{maps FA substructures in each pathway with fold changes from the
“unprocessed_data_test” result and extracts them}
\usage{
FA_sub_extract(
  char_table,
  FA_substructure,
  unprocessed_data_result,
  exact_FA = "no",
  exo_lipid = NULL
)
}
\arguments{
\item{char_table}{Lipid characteristics table. Output of "build_char_table".}

\item{FA_substructure}{FA substructure table. Output of "FA_sub_transform".}

\item{unprocessed_data_result}{Differential expression for unprocessed
lipidomics data. Output of "unprocessed_data_test".}

\item{exact_FA}{"yes" or "no" to decide if the exact FA identity
(e.g., w3 or w6) is known or not.}

\item{exo_lipid}{An character vector specifying exogenous lipid addition. The
exogenous lipids and their adjacent nodes in FA network will not be
decomposed.}
}
\value{
extracted FA substructures
}
\description{
maps FA substructures in each pathway with fold changes from the
“unprocessed_data_test” result and extracts them
}
