% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_sub_transform.R
\name{FA_sub_transform}
\alias{FA_sub_transform}
\title{Decomposes FAs into FA substructures}
\usage{
FA_sub_transform(FA_network, unprocessed_data_result, unmapped_FA = NULL)
}
\arguments{
\item{FA_network}{Trimmed FA biosynthetic network. Output of "build_FA_net".}

\item{unprocessed_data_result}{Differential expression for unprocessed
lipidomics data. Output of "unprocessed_data_test".}

\item{unmapped_FA}{An character vector allowing users to decide which FAs
should be ignored since some FAs can be mapped to more than one node in FA
network (e.g., 20:4;0 for w6-20:4;0 and w3-20:4;0).}
}
\value{
FA substructure table
}
\description{
Decomposes FAs into FA substructures based on the FA biosynthetic network.
}
