% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_substructure_analysis.R
\name{FA_substructure_analysis}
\alias{FA_substructure_analysis}
\title{Website function for fatty acid substructure analysis}
\usage{
FA_substructure_analysis(
  exp_raw,
  method,
  ctrl,
  exp,
  unmapped_FA = c("w9-18:2;0", "w3-20:4;0"),
  exo_lipid = "w3-22:6;0",
  species = "rat",
  add_reaction = NULL,
  delete_reaction = NULL
)
}
\arguments{
\item{exp_raw}{raw expression file}

\item{method}{analysis method}

\item{ctrl}{range of values for control (1:7)}

\item{exp}{range of values for exp (8:13)}

\item{unmapped_FA}{vector of lowly expressed fatty acid isomers to exclude}

\item{exo_lipid}{vector of fatty acids selected to prevent substructure decomposition}

\item{species}{the analysis species, (human, mouse, rat)}

\item{add_reaction}{add a reaction}

\item{delete_reaction}{delete a reaction}
}
\value{
list of tables and plots relevant to the analysis
}
\description{
Website function for fatty acid substructure analysis
}
