% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLipidome-package.R
\docType{package}
\name{iLipidome-package}
\alias{iLipidome}
\alias{iLipidome-package}
\title{iLipidome: Lipidomics Data Analysis In The Context Of The Lipid Biosynthetic Network}
\description{
Numerous biological processes and diseases are influenced by lipid composition. Advances in lipidomics are elucidating their roles, but analyzing and interpreting lipidomics data at the systems level remain challenging. To address this, we present iLipidome, a method for analyzing lipidomics data in the context of the lipid biosynthetic network, thus accounting for the interdependence of measured lipids.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/LewisLabUCSD/iLipidome-package/issues}
}

}
\author{
\strong{Maintainer}: Wen-Jen Lin \email{u104001424@cmu.edu.tw}

Authors:
\itemize{
  \item Austin W.T. Chiang \email{austin.chiang@gmail.com}
  \item Chenguang Liang
  \item Wen-Lung Ma
  \item Wei-Chung Cheng
  \item Nathan E. Lewis \email{n4lewis@ucsd.edu}
}

}
\keyword{internal}
