% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_class_substructure_analysis.R
\name{lipid_class_substructure_analysis}
\alias{lipid_class_substructure_analysis}
\title{Website function for lipid class substructure analysis}
\usage{
lipid_class_substructure_analysis(
  exp_raw,
  method,
  ctrl,
  exp,
  exo_lipid = NULL,
  species = "rat",
  add_reaction = NULL,
  delete_reaction = NULL
)
}
\arguments{
\item{exp_raw}{raw expression input file}

\item{method}{analysis method}

\item{ctrl}{range of values for control (1:7)}

\item{exp}{range of values for exp (8:13)}

\item{exo_lipid}{vector of fatty acids selected to prevent substructure decomposition}

\item{species}{the analysis species, (human, mouse, rat)}

\item{add_reaction}{add a reaction}

\item{delete_reaction}{delete a reaction}
}
\value{
list of tables and plots relevant to the analysis
}
\description{
Website function for lipid class substructure analysis
}
