% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_sub_matrix.R
\name{lipid_sub_matrix}
\alias{lipid_sub_matrix}
\title{Converts expression of FAs to expression of FA substructures.}
\usage{
lipid_sub_matrix(exp_data, sub_data, sub_type = "species")
}
\arguments{
\item{exp_data}{Lipid expression table. Output of "build_char_table".}

\item{sub_data}{Extracted substructure data. Output of "FA_sub_extract".}

\item{sub_type}{"FA", "Class", or "Species" to indicate which type of
substructure.}
}
\value{
FA substructure matrix encoding the frequency of each substructure
}
\description{
Converts expression of FAs to expression of FA substructures.
}
