% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_substructure}
\alias{lipid_substructure}
\title{Dataset Lipid substructure}
\format{
dataframe with 108 rows and 13 variables
\describe{
\item{Lipid}{reference lipid class}
\item{Unit1}{substructure unit 1}
\item{Unit2}{substructure unit 2}
\item{Unit3}{substructure unit 3}
\item{Unit4}{substructure unit 4}
\item{Unit5}{substructure unit 5}
\item{Unit6}{substructure unit 6}
\item{Unit7}{substructure unit 7}
\item{Unit8}{substructure unit 8}
\item{Unit9}{substructure unit 9}
\item{Unit10}{substructure unit 10}
\item{Unit11}{substructure unit 11}
\item{Unit12}{substructure unit 12}
}
}
\source{
{Created in house}
}
\usage{
lipid_substructure
}
\description{
Contains the substructure decomposition rules based on reference lipid class network
}
\examples{
data(lipid_substructure)
}
\keyword{datasets}
