% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipidmaps_sum}
\alias{lipidmaps_sum}
\title{Dataset LipidMaps mapping data}
\format{
dataframe with 16334 rows and 4 variables
\describe{
\item{name}{lipid name}
\item{sub_class_id}{LipidMaps classification ID of lipid}
\item{FA_sum}{summed fatty acid composition of lipid}
\item{FA_split}{individual fatty acid composition of lipid}
}
}
\source{
{https://www.lipidmaps.org/databases/lmsd/download}
}
\usage{
lipidmaps_sum
}
\description{
Contains LipidMaps mapping data for lipids without an exact fatty acid identity
}
\examples{
data(lipidmaps_sum)
}
\keyword{datasets}
