% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLCL_mapping}
\alias{MLCL_mapping}
\title{Dataset LipidMaps mapping data for cardiolipin}
\format{
dataframe with 10102 rows and 8 variables
\describe{
\item{name}{lipid name}
\item{sub_class_id}{LipidMaps classification ID of lipid}
\item{FA_sum}{summed fatty acid of lipid}
\item{FA_split}{individual fatty acid of lipid}
\item{possible_MLCL_sum}{summed fatty acid composition of matched monocardiolipin}
\item{possible_MLCL_split}{individual fatty acid composition of matched monocardiolipin}
\item{possible_PG_sum}{summed fatty acid composition of matched phosphatidylglycerol}
\item{possible_PG_split}{individual fatty acid composition of matched phosphatidylglycerol}
}
}
\source{
{https://www.lipidmaps.org/databases/lmsd/download}
}
\usage{
MLCL_mapping
}
\description{
Contains LipidMaps mapping data for cardiolipin without an exact fatty acid identity
}
\examples{
data(MLCL_mapping)
}
\keyword{datasets}
