% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_edge}
\alias{network_edge}
\title{Dataset Reference lipid class network edge}
\format{
dataframe with 99 rows and 3 variables
\describe{
\item{S1}{lipid class substract}
\item{P1}{lipid class product}
\item{reverse}{A binary indicator to specify if the reaction is reversible or not}
}
}
\source{
{Created in house}
}
\usage{
network_edge
}
\description{
Contains the reference lipid class network edge  data
}
\examples{
data(network_edge)
}
\keyword{datasets}
