% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_node}
\alias{network_node}
\title{Dataset Reference lipid class network node}
\format{
dataframe with 64 rows and 6 variables
\describe{
\item{Abbreviation}{abbreviation of lipid class}
\item{Full_name}{Full name of lipid class}
\item{Class}{category of lipid class}
\item{LIPIDMAPS_subclass}{LipidMaps classification of lipid class}
\item{LIPIDMAPS_subclass_id}{LipidMaps classification ID of lipid class}
\item{FA}{Fatty acid number of lipid class}
}
}
\source{
{Created in house}
}
\usage{
network_node
}
\description{
Contains the reference lipid class network node data
}
\examples{
data(network_node)
}
\keyword{datasets}
