% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_scoring.R
\name{path_scoring}
\alias{path_scoring}
\title{uses substructures to score pathways in biosynthetic network}
\usage{
path_scoring(network, sub_t, calibrate = T, data_type = "FA")
}
\arguments{
\item{network}{Trimmed FA biosynthetic network. Output of "build_FA_net".}

\item{sub_t}{Statistical result table for substructures. Output of "t_test".}

\item{calibrate}{"True" or "False" to calibrate pathway scores or not.}

\item{data_type}{"FA", "Class", or "Species" to indicate which type of
substructure.}
}
\value{
pathway scoring result table
}
\description{
uses substructures to score pathways in biosynthetic network
}
