% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reaction_gene_mapping}
\alias{reaction_gene_mapping}
\title{Dataset Reaction gene}
\format{
dataframe with 390 rows and 3 variables
\describe{
\item{Reaction}{lipid reaction}
\item{species}{species name}
\item{gene}{involved gene name}
}
}
\source{
{https://www.lipidmaps.org/resources/tools/reactions}
}
\usage{
reaction_gene_mapping
}
\description{
Contains lipid reactions and their involved genes
}
\examples{
data(reaction_gene_mapping)
}
\keyword{datasets}
