% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_scoring.R
\name{reaction_scoring}
\alias{reaction_scoring}
\title{evaluates each reaction in biosynthetic network using substructures}
\usage{
reaction_scoring(
  network,
  sub_exp,
  sub_t,
  ctrl = 1:7,
  exp = 8:13,
  stat = "p",
  Species = "human"
)
}
\arguments{
\item{network}{Trimmed FA biosynthetic network. Output of "build_FA_net".}

\item{sub_exp}{Substructure profile. Output of "lipid_sub_matrix".}

\item{sub_t}{Statistical result table for substructures. Output of "t_test".}

\item{ctrl}{An integer vector specifying samples in the control group in
lipid expression table. Note that the first column containing lipid names is
not counted.}

\item{exp}{An integer vector specifying samples in the experimental group in
lipid expression table. Note that the first column containing lipid names is
not counted.}

\item{Species}{"human", "mouse", or "rat" can be used to label
species-specific genes for lipid reactions.}
}
\value{
reaction scoring result table
}
\description{
evaluates each reaction in biosynthetic network using substructures
}
