% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_sub_transform.R
\name{species_sub_transform}
\alias{species_sub_transform}
\title{decomposes lipids into species substructures based on the lipid biosynthetic
network.}
\usage{
species_sub_transform(char, lipid_substructure, network_node)
}
\arguments{
\item{char}{Lipid characteristics table. Output of Output of
"build_char_table".}

\item{lipid_substructure}{A data.frame recording lipid class substructures
computed from the reference lipid biosynthetic network in iLipdiome.}

\item{network_node}{A data.frame recording iLipidome-supported lipid classes
and related information.}
}
\value{
lipid species substructure table
}
\description{
decomposes lipids into species substructures based on the lipid biosynthetic
network.
}
