% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{t-test}
\usage{
t_test(data, ctrl, exp, method = "t.test", significant = "p_value")
}
\arguments{
\item{data}{FA substructure profile. Output of "lipid_sub_matrix".}

\item{ctrl}{An integer vector specifying samples in the control group in
lipid expression table. Note that the first column containing lipid names
is not counted.}

\item{exp}{An integer vector specifying samples in the experimental group in
lipid expression table. Note that the first column containing lipid names is
not counted.}

\item{method}{A character string indicating which statistical method to be.
Currently, iLipidome supports two-sample t-tests ("t.test"), Wilcoxon tests
("wilcox.test"), or moderated t-tests ("mod.t.test").}

\item{significant}{"p_value" or "adj_p_value" can be used for selecting
significant lipid features.}
}
\value{
statistical results table
}
\description{
t-test
}
