% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unprocessed_data_test.R
\name{unprocessed_data_test}
\alias{unprocessed_data_test}
\title{Uses output of "build_char_table" to perform differential expression for
three types of data: 1.) lipid species, 2.) fatty acids, and 3.) lipid
classes}
\usage{
unprocessed_data_test(
  exp_data,
  char_table,
  method = "t.test",
  significant = p_value,
  ctrl_group,
  exp_group
)
}
\arguments{
\item{exp_data}{Lipid expression table. Output of "build_char_table".}

\item{char_table}{Lipid characteristics table. Output of "build_char_table".}

\item{method}{A character string indicating which statistical method to be.
Currently, iLipidome supports two-sample t-tests ("t.test"), Wilcoxon tests
("wilcox.test"), or moderated t-tests ("mod.t.test").}

\item{significant}{"p_value" or "adj_p_value" can be used for selecting
significant lipid features.}

\item{ctrl_group}{An integer vector specifying samples in the control group
in lipid expression table. Note that the first column containing lipid names
is not counted.}

\item{exp_group}{An integer vector specifying samples in the experimental
group in lipid expression table. Note that the first column containing lipid
names is not counted.}
}
\value{
list of expression tables
}
\description{
Uses output of "build_char_table" to perform differential expression for
three types of data: 1.) lipid species, 2.) fatty acids, and 3.) lipid
classes
}
