% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_species_net.R
\name{build_species_net}
\alias{build_species_net}
\title{uses species substructures to contruct lipid biosynthetic network}
\usage{
build_species_net(species_substructure)
}
\arguments{
\item{species_substructure}{Output of "species_sub_transform"}
}
\value{
lipid species biosynthetic network
}
\description{
uses species substructures to contruct lipid biosynthetic network
}
