% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_network.R
\name{draw_network}
\alias{draw_network}
\title{build biosynthetic network using substructures, pathway and reaction scoring
results}
\usage{
draw_network(
  network_data,
  DE_data,
  if_species = F,
  significant = "p_value",
  path_scoring_result,
  reaction_scoring_result,
  top_n = 3,
  path_type = "both"
)
}
\arguments{
\item{network_data}{Trimmed FA biosynthetic network. Output of "build_FA_net"}

\item{DE_data}{Statistical result table for substructures. Output of "t_test"}

\item{if_species}{"True" or "False" to indicate data type is lipid species or
the others (FA or class).}

\item{significant}{"p_value" or "adj_p_value" can be used for selecting
significant lipid features.}

\item{path_scoring_result}{Pathway scoring result table. Output of
"path_scoring".}

\item{top_n}{1 to 5 to label top N pathways and reactions.}

\item{path_type}{"active", "suppressed", or "both" is used to select specific
types of top N pathways and reactions.}

\item{reaction_scoring_resul:}{Reaction scoring result table. Output of
"reaction_scoring".}
}
\value{
biosynthetic network node
}
\description{
build biosynthetic network using substructures, pathway and reaction scoring
results
}
