% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FA_network}
\alias{FA_network}
\title{Dataset Reference fatty acid network}
\format{
A dataframe with 51 rows and 5 variables
\describe{
\item{S1}{fatty acid substrate}
\item{P1}{fatty acid product}
\item{S1_detail}{fatty acid substrate with double bond specification}
\item{P1_detail}{fatty acid product with double bond specification}
\item{pathway}{fatty acid pathway}
}
}
\source{
{Created in-house}
}
\usage{
FA_network
}
\description{
Contains the reference fatty acid  network data
}
\examples{
data(FA_network)
}
\keyword{datasets}
